[#312] Add a maven-central-publish plugin to manage publication of multiple artifacts
