+++ buildSrc/src/main/kotlin/com/orange/ods/gradle/MavenCentralPublishPluginExtension.kt
@@ -1,0 +1,20 @@
+/*
+ *
+ *  Copyright 2021 Orange
+ *
+ *  Use of this source code is governed by an MIT-style
+ *  license that can be found in the LICENSE file or at
+ *  https://opensource.org/licenses/MIT.
+ * /
+ */
+
+package com.orange.ods.gradle
+
+abstract class MavenCentralPublishPluginExtension {
+
+    companion object {
+        const val Name = "mavenCentralPublish"
+    }
+
+    var artifactId: String? = null
+}
