+++ theme-contract/build.gradle.kts
@@ -13,6 +13,7 @@ import com.orange.ods.gradle.Versions
 plugins {
     id("com.android.library")
     id("org.jetbrains.kotlin.android")
+    id("maven-central-publish")
 }
 
 android {
