+++ theme-innovation-cup/build.gradle.kts
@@ -17,6 +17,7 @@ plugins {
     id("kotlin-kapt")
     id("dagger.hilt.android.plugin")
     id("kotlin-parcelize")
+    id("maven-central-publish")
 }
 
 android {
